requirements.txt
# Monitor dependencies
scout-apm==2.26.1
sentry-sdk==1.16.0

# Production dependencies
Babel==2.11.0
bleach==5.0.0
boto3==1.26.57
celery==5.2.7
click==8.1.3
dj-database-url==0.5.0
django-anymail==9.0
django-basic-auth-ip-whitelist==0.5
django-bleach==3.0.1
django-countries==7.5
django-elevate==2.0.3
django-extensions==3.2.1
django-file-form==3.4.3
django-filter==2.4.0
django-formtools==2.4.1
django-fsm==2.8.1
django-heroku==0.3.1
django-hijack==3.2.6
django-htmx==1.13.0
django-pagedown==2.2.1
# django-pwned-passwords==4.1.0
https://github.com/slinkymanbyday/django-pwned-passwords/archive/58c7b832df7360a21fd8edeaaf9f897c7517baf1.zip
django-ratelimit==4.0.0
django-referrer-policy==1.0
django-select2==8.0.0
django-slack==5.18.0
django-storages==1.13.2
django-tables2==2.5.1
django-tinymce==3.5.0
django-two-factor-auth==1.14.0
django-web-components==0.1.1
django==4.1.8
djangorestframework-api-key==2.3.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-yasg==1.21.4
environs==9.5.0
gunicorn==20.1.0
heroicons==2.2.0
python-docx<1.0.0
htmldocx==0.0.6
lark==1.1.5
mailchimp3==3.0.17
mistune==2.0.4
more-itertools==9.0.0
phonenumberslite==8.13.4
Pillow==9.4.0
psycopg2-binary
qrcode==7.4.2
reportlab==3.6.13
social_auth_app_django==5.0.0
tablib==3.3.0
tomd==0.1.3
wagtail-cache==2.2.0
wagtail-purge==0.3.0
wagtail==4.2.2
whitenoise==5.3.0
xhtml2pdf==0.2.8
xmltodict==0.13.0
